migration change

+++ database/migrations/2021_05_14_113023_create_offers_table.php
@@ -21,7 +21,7 @@ class CreateOffersTable extends Migration
             $table->string('region');
             $table->string('ville');
             $table->string('image')->nullable();
-            $table->BigInteger('user_id')->unsigned();
+            $table->integer('user_id')->unsigned();
             $table->string('slug')->nullable();
             $table->timestamps();
